+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Resolver
 
+### Metroid Prime
+
+- Added: FAQ Entry about non-Superheated rooms
+
 - Fixed: Some cases of seeds being wrongly considered as impossible.
 
 ## [7.5.1] - 2024-04-??